+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - added alternative dice notation
 - added outputs to result
+- added formatter to result output
 
 ## [0.0.8]
 ### Added